requirements.txt
discord.py
aiocron
aiocron